pyproject.toml
[project]
name = "mcp-viewer"
version = "0.1.0"
description = "Document Viewer - Interactive ALTO XML visualization"
requires-python = ">=3.13"
dependencies = [
    "httpx[http2]>=0.28.0",
    "fastmcp>=3.0.2",
    "pillow>=12.1.1",
    "py-key-value-aio[memory]>=0.4.4",
]

[project.scripts]
mcp-viewer = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
include = ["server.py", "src/**", "dist/**"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "respx>=0.22.0",
]
